pyproject.toml
[project]
name = "mtihani-api"
version = "0.1.0"
description = ""
authors = [
    {name = "melaniefayne",email = "melaniefayne33@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "semantic-kernel (>=1.28.1,<2.0.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "accelerate (>=1.6.0,<2.0.0)",
    "huggingface-hub (>=0.30.2,<0.31.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
